benchmark: split TTS HTTP client contracts
